+++ Mailer/migrations/20201209084222_mail_template_search_indexes.php
@@ -1,0 +1,16 @@
+<?php
+declare(strict_types=1);
+
+use Phinx\Migration\AbstractMigration;
+
+final class MailTemplateSearchIndexes extends AbstractMigration
+{
+    public function change(): void
+    {
+        $this->table('mail_templates')
+            ->addIndex('name')
+            ->addIndex('subject')
+            ->addIndex('description')
+            ->update();
+    }
+}