+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 - Fixed issues with settings page caused by internal config names renaming. remp/crm#1616
 - Fixed main search issue due to the internal changes in the past release. remp/remp#786
+- Improved search speed of the main search bar. remp/remp#786
 
 ### [Beam]